# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "樱花动漫后台服务 API（由 PHP/Phalcon 迁移至 Go/Gin）。",
        "title": "yhdm_service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/backend",
    "paths": {
...
        "/app/article-categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "文章分类列表(分页)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称模糊",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/yhdm_service_internal_service.ArticleCategoryListResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/article-detail": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yhdm_service_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/app/articles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "文章列表(分页)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标题模糊",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/yhdm_service_internal_service.ArticleListResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/block-positions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "模块位置列表(分页)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称模糊",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/yhdm_service_internal_service.BlockPositionListResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/delete-article": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "description": "文章ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.idReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yhdm_service_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/app/delete-article-category": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章分类",
                "parameters": [
                    {
                        "description": "分类ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.idReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yhdm_service_internal_response.Body"
                        }
                    }
                }
            }
        },
...
        "/app/delete-block-position": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除模块位置",
                "parameters": [
                    {
                        "description": "模块位置ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.idReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yhdm_service_internal_response.Body"
                        }
                    }
                }
            }
        },
...
        "/app/save-article": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "新增/更新文章(id\u003e0 为更新)",
                "parameters": [
                    {
                        "description": "文章内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.saveArticleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {id, message}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/app/save-article-category": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "新增/更新文章分类(id\u003e0 为更新)",
                "parameters": [
                    {
                        "description": "分类信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.saveCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {id, message}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/app/save-block-position": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "新增/更新模块位置(id\u003e0 为更新)",
                "parameters": [
                    {
                        "description": "模块位置信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.saveBlockReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {id, message}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/yhdm_service_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_admin.idReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_admin.saveArticleReq": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "click": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "is_recommend": {
                    "type": "integer"
                },
                "seo_description": {
                    "type": "string"
                },
                "seo_keywords": {
                    "type": "string"
                },
                "show_dialog": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_admin.saveBlockReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_admin.saveCategoryReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
        "yhdm_service_internal_model.Article": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "click": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "is_recommend": {
                    "type": "integer"
                },
                "seo_description": {
                    "type": "string"
                },
                "seo_keywords": {
                    "type": "string"
                },
                "show_dialog": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "yhdm_service_internal_model.ArticleCategory": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "yhdm_service_internal_model.BlockPosition": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "yhdm_service_internal_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "yhdm_service_internal_service.ArticleCategoryListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/yhdm_service_internal_model.ArticleCategory"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "yhdm_service_internal_service.ArticleListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/yhdm_service_internal_model.Article"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "yhdm_service_internal_service.BlockPositionListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/yhdm_service_internal_model.BlockPosition"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}